requirements.txt
astroid==1.6.0
certifi==2017.11.5
chardet==3.0.4
ChatterBot==0.8.1
chatterbot-corpus==1.1.1
dateparser==0.6.0
Django==2.0
djangorestframework==3.7.7
djangotoolbox==1.8.0
future==0.16.0
idna==2.6
isort==4.2.15
jsondatabase==0.1.7
lazy-object-proxy==1.3.1
mathparse==0.1.1
mccabe==0.6.1
nltk==3.2.5
oauthlib==2.0.6
pylint==1.7.5
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
pymongo==3.6.0
python-dateutil==2.6.1
python-twitter==3.3
pytz==2017.3
PyYAML==3.12
regex==2017.12.12
requests==2.18.4
requests-oauthlib==0.8.0
ruamel.yaml==0.15.35
six==1.11.0
SQLAlchemy==1.1.15
tzlocal==1.5.1
urllib3==1.22
websocket-client==0.44.0
wrapt==1.10.11
